split job.js into single files
