+++ lib/job/disable.js
@@ -1,0 +1,10 @@
+'use strict';
+
+/**
+ * Prevents the job type from running
+ * @returns {exports} instance of Job
+ */
+module.exports = function() {
+  this.attrs.disabled = true;
+  return this;
+};
