+++ lib/job/enable.js
@@ -1,0 +1,10 @@
+'use strict';
+
+/**
+ * Allows job type to run
+ * @returns {exports} instance of Job
+ */
+module.exports = function() {
+  this.attrs.disabled = false;
+  return this;
+};
