+++ lib/job/repeat-at.js
@@ -1,0 +1,11 @@
+'use strict';
+
+/**
+ * Sets a job to repeat at a specific time
+ * @param {String} time time to repeat job at (human readable or number)
+ * @returns {exports} instance of Job
+ */
+module.exports = function(time) {
+  this.attrs.repeatAt = time;
+  return this;
+};
